refactor: split cmd/kbn/api.go into separate files by API responsibility
- cmd/kbn/api.go → cmd/kbn/api_bootnode.go (BootnodeAPI)
- cmd/kbn/api.go → cmd/kbn/api_bootnode_registry.go (BootnodeRegistryAPI)
